PHOENIX-6978 Redesign Phoenix TTL for Views (#1938)
* PHOENIX-6979 Use HBase TTL as TTL for Tables only at Phoenix Level when phoenix.table.ttl.enabled is true (#1622)

* PHOENIX-6979 (Addendum) Use HBase TTL as TTL for Tables only at Phoenix Level when phoenix.table.ttl.enabled is true (#1670)

* PHOENIX-7022 :- Add new columns TTL and ROW_KEY_PREFIX (#1660)

* PHOENIX-7022 :- Add TTL and ROWKEY_PREFIX Column and Use TTL instead of PHOENIX_TTL for storing TTL in SYSCAT


* PHOENIX-7023 : Deprecate columns PHOENIX_TTL and PHOENIX_TTL_HWM and related code (#1688)

* Deprecate columns PHOENIX_TTL and PHOENIX_TTL_HWM and related code


* PHOENIX-7041 Populate ROW_KEY_PREFIX column when creating views (#1733)

PHOENIX-7041 Populate ROW_KEY_PREFIX column when creating views
---------


* PHOENIX-7040 : Support TTL for views using the new column TTL in SYSTEM.CATALOG (#1710)

PHOENIX-7040  Support TTL for views using the new column TTL in SYSTEM.CATALOG
---------


* PHOENIX-7041 (ADDENDUM ): Populate ROW_KEY_PREFIX column when creating views (#1821)

* Refactored ROW_KEY_PREFIX column name to ROW_KEY_MATCHER

* PHOENIX-7040 (ADDENDUM): Support TTL for views using the new column TTL in SYSTEM.CATALOG (#1782)


* PHOENIX-6996 :- Provide an upgrade path for Phoenix tables with HBase TTL to move the… (#1752)

* Provide an upgrade path for Phoenix tables with HBase TTL to move their TTL spec to SYSTEM.CATALOG

* Adding a config to enable/disable ViewTTL feature

* PHOENIX-7041 (ADDENDUM-2): ROW_KEY_MATCHER should support various tenantId data types (#1900)

* PHOENIX-7329 Change TTL column type to VARCHAR in syscat (#1905)

PHOENIX-7329 Changing syscat column to type VARCHAR

* PHOENIX-7108 Provide support for pruning expired rows of views using Phoenix level compactions (#1908)

---------

+++ phoenix-core-server/src/main/java/org/apache/phoenix/coprocessor/GlobalIndexRegionScanner.java
@@ -287,11 +287,11 @@ public abstract class GlobalIndexRegionScanner extends BaseRegionScanner {
         throw new IllegalStateException("No cell found");
     }
 
-    protected static boolean isTimestampBeforeTTL(long tableTTL, long currentTime, long tsToCheck) {
+    protected static boolean isTimestampBeforeTTL(int tableTTL, long currentTime, long tsToCheck) {
         if (tableTTL == HConstants.FOREVER) {
             return false;
         }
-        return tsToCheck < (currentTime - tableTTL * 1000);
+        return tsToCheck < (currentTime - tableTTL * 1000L);
     }
 
     protected static boolean isTimestampBeyondMaxLookBack(long maxLookBackInMills,
